Splitting up EventGraph into partial classes
